Rename 'arqivr' to 'stoker'.
