+++ stoker/cli.py
